+++ app/utils/authOptions.ts
@@ -1,0 +1,24 @@
+import NextAuth from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+import { NextAuthOptions } from "next-auth";
+
+export const authOptions: NextAuthOptions = {
+    providers: [
+      GoogleProvider({
+        clientId: process.env.GOOGLE_CLIENT_ID!,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+      }),
+    ],
+    pages: {
+      signIn: "/",
+    },
+    callbacks: {
+      async signIn({ user, account, profile }) {
+        return true;
+      },
+      async redirect({ url, baseUrl }) {
+        return baseUrl + "/waiting";
+      },
+    },
+    debug: true,
+  };
